+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/testing/scenarios/LoadTestUtil.java
@@ -24,6 +24,7 @@ final class LoadTestUtil {
 
   private static final int MAX_ARTIST_ID = 275;
   private static final int MAX_CUSTOMER_ID = 59;
+  private static final int MAX_TRACK_ID = 3503;
 
   static final Random RANDOM = new Random();
 
@@ -33,5 +34,9 @@ final class LoadTestUtil {
 
   static long randomCustomerId() {
     return RANDOM.nextInt(MAX_CUSTOMER_ID) + 1;
+  }
+
+  static long randomTrackId() {
+    return RANDOM.nextInt(MAX_TRACK_ID) + 1;
   }
 }