+++ changelog.md
@@ -6,3 +6,7 @@ Codion Change Log
 - EntityTable.layoutPanel(), tablePanel parameter replaced with tableComponent, no longer restricted to JPanel.
 - Jasper Reports and Apache Commons Logging upgraded.
 - Logback and Spotless upgraded.
+- EntityServerConfiguration, connectionPoolProvider renamed connectionPoolFactory
+- ConnectionPoolFactory.createConnectionPoolWrapper() renamed createConnectionPool()
+- WaitCursor replaced with Cursors utility class, wait cursor usage reduced and simplified throughout.
+- EntityEditPanel.beforeInsert() and beforeUpdate() no longer throw ValidationException, now called outside of try/catch block along with beforeDelete(). EntityTablePanel.beforeDelete() moved outside of try/catch block, delete() added for deleting without confirmation.